refactor: move native block list to common block list
