+++ morph/react-native/block.js
@@ -4,7 +4,7 @@ import * as BlockExplicitChildren from '../react/block-explicit-children.js'
 // import * as BlockGoTo from './block-go-to.js'
 import * as BlockInList from '../react/block-in-list.js'
 import * as BlockName from './block-name.js'
-import * as BlockList from './block-list.js'
+import * as BlockList from '../react/block-list.js'
 import * as BlockGroup from '../react/block-group.js'
 import * as BlockProperties from './block-properties.js'
 import * as BlockProxy from '../react/block-proxy.js'
