+++ morph/react/block-group.js
@@ -1,5 +1,4 @@
 export function enter(node, parent, state) {
-  debugger
   if (
     node.isGroup &&
     node.children.length > 0 &&
