+++ morph/react/block-list.js
@@ -5,14 +5,23 @@ export function enter(node, parent, state) {
     const from = getProp(node, 'from')
     if (!from) return
 
-    state.render.push(
-      `{Array.isArray(${from.value}) && ${from.value}.map((item, index) => `
-    )
+    if (node.nameFinal === 'FlatList') {
+      state.render.push(`data={${from.value}} renderItem={({ item }) =>`)
+    } else {
+      state.render.push(
+        `{Array.isArray(${from.value}) && ${from.value}.map((item, index) => `
+      )
+    }
   }
 }
 
 export function leave(node, parent, state) {
+  debugger
   if (isList(node)) {
-    state.render.push(')}')
+    if (node.nameFinal === 'FlatList') {
+      state.render.push('}')
+    } else {
+      state.render.push(')}')
+    }
   }
 }
